+++ opendmi/src/format/xml/handlers.c
@@ -341,7 +341,7 @@ bool dmi_xml_entity_attr_value(
         return dmi_xml_entity_attr_set(session, attr, value);
 
     do {
-        text = dmi_attribute_format(attr, value, false);
+        text = dmi_attribute_format(session->context, attr, value, false);
         if (text == nullptr)
             break;